# app/frontend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/verification/{id}": {
      "get": {
        "description": "Retrieve the current status and details of a verification request. Part of v1 API.",
        "operationId": "VerificationController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the verification request",
            "schema": {
              "example": "clv789xyz123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification status retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "clv789xyz123",
                    "userId": "clu456def789",
                    "documentType": "NATIONAL_ID",
                    "status": "APPROVED",
                    "submittedAt": "2025-01-23T11:00:00.000Z",
                    "reviewedAt": "2025-01-23T14:30:00.000Z",
                    "verificationCode": "VER-2025-0123",
                    "notes": "All documents verified successfully"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - valid JWT token required."
          },
          "404": {
            "description": "The specified verification request was not found."
          }
        },
        "security": [
          {
            "x-api-key": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get verification status (v1)",
        "tags": [
          "Verification"
        ]
      }
    },
...
    "/api/v1/verification/{id}/notes": {
      "post": {
        "description": "Adds a secure internal note for staff review only.",
        "operationId": "VerificationController_addNote_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInternalNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Internal note added successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalNoteResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Access denied - staff role required."
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "Add an internal note to a verification record",
        "tags": [
          "Verification"
        ]
      },
      "get": {
        "description": "Retrieves all internal notes for a specific verification.",
        "operationId": "VerificationController_getNotes_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Internal notes retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InternalNoteResponseDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access denied - staff role required."
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "List internal notes for a verification record",
        "tags": [
          "Verification"
        ]
      }
    },
    "/api/v1/verification-inbox": {
      "get": {
        "description": "Retrieve verification requests with filtering and pagination. Shows pending_review, approved, rejected, and needs_resubmission states.",
        "operationId": "VerificationInboxController_getInbox_v1",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by verification status",
            "schema": {
              "enum": [
                "pending_review",
                "approved",
                "rejected",
                "needs_resubmission"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 20)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification inbox retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "items": [
                      {
                        "id": "clv789xyz123",
                        "status": "pending_review",
                        "createdAt": "2025-01-23T11:00:00.000Z",
                        "reviewedAt": null,
                        "nextStepMessage": "Review identity documents for authenticity",
                        "deepLink": "/verification/clv789xyz123"
                      }
                    ],
                    "total": 45,
                    "page": 1,
                    "limit": 20,
                    "totalPages": 3
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - valid JWT token required."
          },
          "403": {
            "description": "Access denied - operator or admin role required."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get verification inbox",
        "tags": [
          "Verification Inbox"
        ]
      }
    },
    "/api/v1/verification-inbox/stats": {
      "get": {
        "description": "Retrieve counts of verification requests by status for dashboard display.",
        "operationId": "VerificationInboxController_getStats_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "pending_review": 15,
                    "approved": 120,
                    "rejected": 8,
                    "needs_resubmission": 5,
                    "total": 148
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - valid JWT token required."
          },
          "403": {
            "description": "Access denied - operator or admin role required."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get verification inbox statistics",
        "tags": [
          "Verification Inbox"
        ]
      }
    },
    "/api/v1/verification-inbox/{id}/approve": {
      "post": {
        "description": "Mark a verification request as approved and set next step messaging.",
        "operationId": "VerificationInboxController_approve_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the verification request",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nextStepMessage": {
                    "type": "string"
                  },
                  "internalNote": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification approved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "clv789xyz123",
                    "status": "approved",
                    "reviewedAt": "2025-01-23T14:30:00.000Z",
                    "nextStepMessage": "Verification approved. Proceed to disbursement."
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or verification already processed."
          },
          "404": {
            "description": "The specified verification request was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Approve verification request",
        "tags": [
          "Verification Inbox"
        ]
      }
    },
    "/api/v1/verification-inbox/{id}/reject": {
      "post": {
        "description": "Mark a verification request as rejected with reason and next step messaging.",
        "operationId": "VerificationInboxController_reject_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the verification request",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "rejectionReason"
                ],
                "properties": {
                  "rejectionReason": {
                    "type": "string"
                  },
                  "nextStepMessage": {
                    "type": "string"
                  },
                  "internalNote": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification rejected successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "clv789xyz123",
                    "status": "rejected",
                    "reviewedAt": "2025-01-23T14:30:00.000Z",
                    "rejectionReason": "Document appears fraudulent",
                    "nextStepMessage": "Please resubmit with valid documentation"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or verification already processed."
          },
          "404": {
            "description": "The specified verification request was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Reject verification request",
        "tags": [
          "Verification Inbox"
        ]
      }
    },
    "/api/v1/verification-inbox/{id}/request-resubmission": {
      "post": {
        "description": "Mark a verification request as needing resubmission with specific requirements.",
        "operationId": "VerificationInboxController_requestResubmission_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the verification request",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "rejectionReason",
                  "nextStepMessage"
                ],
                "properties": {
                  "rejectionReason": {
                    "type": "string"
                  },
                  "nextStepMessage": {
                    "type": "string"
                  },
                  "internalNote": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resubmission requested successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "clv789xyz123",
                    "status": "needs_resubmission",
                    "reviewedAt": "2025-01-23T14:30:00.000Z",
                    "rejectionReason": "Document expired",
                    "nextStepMessage": "Please submit a current government-issued ID"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or verification already processed."
          },
          "404": {
            "description": "The specified verification request was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Request resubmission for verification",
        "tags": [
          "Verification Inbox"
        ]
      }
    },
    "/api/v1/verification-inbox/{id}": {
      "get": {
        "description": "Retrieve detailed information about a specific verification request including review history.",
        "operationId": "VerificationInboxController_getDetails_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the verification request",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification details retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "clv789xyz123",
                    "status": "pending_review",
                    "createdAt": "2025-01-23T11:00:00.000Z",
                    "reviewedAt": null,
                    "reviewedBy": null,
                    "rejectionReason": null,
                    "nextStepMessage": "Review identity documents for authenticity",
                    "deepLink": "/verification/clv789xyz123"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - valid JWT token required."
          },
          "404": {
            "description": "The specified verification request was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get verification request details",
        "tags": [
          "Verification Inbox"
        ]
      }
    },
    "/api/v1/verification-inbox/{id}/notes": {
      "post": {
        "description": "Add an internal note visible only to reviewers. The action is recorded in the audit trail.",
        "operationId": "VerificationInboxController_addInternalNote_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the verification request",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Internal note added successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "note-abc123",
                    "entityType": "verification",
                    "entityId": "clv789xyz123",
                    "content": "Contacted applicant for additional documents.",
                    "authorId": "reviewer-001",
                    "category": "follow_up",
                    "createdAt": "2025-01-23T15:00:00.000Z"
                  }
                }
              }
            }
          },
          "404": {
            "description": "The specified verification request was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Add internal note to verification request",
        "tags": [
          "Verification Inbox"
        ]
      },
      "get": {
        "description": "Retrieve all internal notes attached to a verification request.",
        "operationId": "VerificationInboxController_getInternalNotes_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the verification request",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Internal notes retrieved successfully."
          },
          "404": {
            "description": "The specified verification request was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List internal notes for a verification request",
        "tags": [
          "Verification Inbox"
        ]
      }
    },
...
    "/api/v1/entity-linking/verification/{verificationId}": {
      "get": {
        "description": "Retrieve all entity links associated with a specific verification",
        "operationId": "EntityLinkingController_getLinksByVerification_v1",
        "parameters": [
          {
            "name": "verificationId",
            "required": true,
            "in": "path",
            "description": "Verification ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "organization",
                "location",
                "asset",
                "project"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get entity links by verification",
        "tags": [
          "Entity Linking"
        ]
      }
    },
...
  },
  "info": {
    "title": "ChainForge API",
    "description": "API documentation for ChainForge platform - Humanitarian aid and verification system\n\n## API Versioning\n\nThis API uses URI-based versioning. The current version is **v1**.\n\n### Version Format\nAll endpoints are prefixed with the version number: `/api/v1/...`\n\n### Supported Versions\n| Version | Status | Description |\n|---------|--------|-------------|\n| v1 | Current | Active version with full support |\n\n### Deprecation Policy\n- Deprecated endpoints will be marked with `@Deprecated` in the documentation\n- Deprecated versions will be supported for at least 6 months after deprecation notice\n- Clients will receive deprecation warnings via the `Sunset` HTTP header\n- Migration guides will be provided for major version changes\n\n### Future Versions\nWhen new versions are released:\n- New endpoints will be available at `/api/v2/...`, etc.\n- Previous versions remain accessible during the deprecation period\n- Clients should monitor the API documentation for version updates",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    },
    {
      "url": "https://api.chainforge.app",
      "description": "Staging"
    },
    {
      "url": "https://api.chainforge.app",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "description": "Enter JWT token"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for external access"
      }
    },
    "schemas": {
...
      "CreateInternalNoteDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "The content of the internal note.",
            "example": "Investigation shows that the recipient has provided valid evidence."
          },
          "category": {
            "type": "string",
            "description": "Optional category for the note.",
            "example": "investigation"
          }
        },
        "required": [
          "content"
        ]
      },
      "InternalNoteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clv789xyz123"
          },
          "entityType": {
            "type": "string",
            "example": "claim"
          },
          "entityId": {
            "type": "string",
            "example": "clv789xyz123"
          },
          "content": {
            "type": "string",
            "example": "Investigation notes..."
          },
          "authorId": {
            "type": "string",
            "example": "clv789xyz123"
          },
          "category": {
            "type": "object",
            "example": "investigation",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-23T11:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-23T11:00:00.000Z"
          }
        },
        "required": [
          "id",
          "entityType",
          "entityId",
          "content",
          "authorId",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}